reorganize and get rid of unused function & corresponding test
