+++ internal/trello/trello_test.go
@@ -132,60 +132,3 @@ func TestNewClient(t *testing.T) {
 		})
 	}
 }
-
-func TestContains(t *testing.T) {
-	str := "placeholder"
-
-	tt := []struct {
-		name    string
-		list    []string
-		query   string
-		outcome bool
-	}{
-		{
-			name:    "all empty",
-			list:    []string{},
-			query:   "",
-			outcome: false,
-		},
-		{
-			name:    "empty query",
-			list:    []string{str},
-			query:   "",
-			outcome: false,
-		},
-		{
-			name:    "empty list",
-			list:    []string{},
-			query:   str,
-			outcome: false,
-		},
-		{
-			name:    "no match",
-			list:    []string{"a", "b", "c"},
-			query:   "d",
-			outcome: false,
-		},
-		{
-			name:    "match",
-			list:    []string{"a", "b", "c", str},
-			query:   str,
-			outcome: true,
-		},
-	}
-
-	for _, tc := range tt {
-		t.Run(tc.name, func(t *testing.T) {
-			outcome := contains(tc.list, tc.query)
-			if outcome == tc.outcome {
-				return
-			}
-
-			prefix := "expected "
-			if !tc.outcome {
-				prefix = "did not expect "
-			}
-			t.Fatalf("%s %s to be in the list %v, got %t", prefix, tc.query, tc.list, outcome)
-		})
-	}
-}
